test(workers): split decision envelope runner tests
